Edit : Terms
Pages에서 componedts로 이동
